// components/calendar/types.ts
export type Time = {
  hour: string
  minute: string
}
// ...
export interface TimepickerProps {
  onChange: (selectedTime: Time) => void
  value?: string
  hoursLabel?: string
  minutesLabel?: string
  secondsLabel?: string
  placeholder?: string
  label?: string
  nowButtonLabel?: string
}